Return the maximum privilege in open session command
This commit does the following:

- Set the maximum privilege role in the session instead of current privilege
  in the implementation of the open session request/response.

- In open session response, return the maximum privilege of the session
  instead of the current privilege level.

- Update RAKP12 and RAKP34 implementation

Tested: Checked the session setup works fine with ipmitool and freeipmi
utility ipmipower.

+++ command/rakp34.cpp
@@ -126,7 +126,7 @@ std::vector<uint8_t> RAKP34(const std::vector<uint8_t>& inPayload,
     auto rcSessionID = endian::to_ipmi(session->getRCSessionID());
 
     // Session Privilege Level
-    auto sessPrivLevel = session->reqMaxPrivLevel;
+    auto sessPrivLevel = static_cast<uint8_t>(session->reqMaxPrivLevel);
 
     // User Name Length Byte
     auto userLength = static_cast<uint8_t>(session->userName.size());
